Chore: reorganize folder structure( platform enum)
